Part 3. Erik Ramsey
Used my own functions, but most likely can be removed once project is all put together. Was able to find a way to see when the given piece returned to its original position. Only piece not done is the king as I was getting an error with the variation of the selector function and I could not figure out how to fix it. Also, didn't want to overwrite code, but I imagine you could also run the replicate function to get the number of turns it first took to return to the original position multiple times and then plot a distribution from there. If you have an idea to fix the king function or if yours works (Kirk) feel free to change stuff if you look at it before we meet tomorrow.