Cleaned up Part 3.
I was able to write a function that produced a density plot for the number of times a piece returned to its original position in 10,000 moves. I called the function 5 times for each piece to get five distribution plots. If we are supposed to knit this to either an html or a word doc, I'm not sure if we want 25 distribution plots for this part or to just include one plot for one position for each piece and keep the other positions hidden from the knitted document, but visible in the markdown. Let me know either way and I can change it or whoever ends up compiling the whole thing can change it. The chunks also take like two minutes each to run. Not sure if it's just my computer or what as it is quite old now. The only piece I don't have done is the king as I haven't been able to figure out how to get the movement function to work for it either. Let me know if anyone figures it out if it's not already in the group chat.