Part 3 without King
I think this is the finalized code other than the king. Being that that mover function is different I didn't get to it before I stopped for the night. Like I said in the group chat I am going to try and finish that before I have work in the late afternoon. I reduced the number of turns from 10,000 to 5000 as I was still getting good results. I could probably drop it to around 3000 and still be getting around 40 observations for when the piece returns to its original positions. I was also able to get all 5 starting positions for the 4 pieces other than the king on one plot. If you are reading through descriptions I wrote for each of the distributions feel free to change things.