Final? Code for Part 3.
I think this should be all of the code to plot the distributions of 5 different starting positions for the 5 pieces. I know it looks way different from the previous code, but that is because I had to do a bunch of data frame manipulation with the tidyverse in order to get the 5 different starter positions all on the same plot. Let me know if you think I should comment out some code. I was pretty tired working on it, so I got a bit lazy with that portion. I figured I could go back and comment later. I also don't think we really need to include the part where I sort of compared the distributions of the 5 pieces in the final paragraph I wrote. I don't know how much it actually helps explain anything and it's not asked in the question. I sort of just added it just in case. Let me know either way. I may be sleeping when you see this pull request.